Keys, Glasses, Bandanas, Cowboy Hats
All are sized at 48x48 pixels.
Intended use: Use sprite 0 when the item is not obtained, use sprite 1 when item is obtained.
The Key Piece and Key New Home items are weird because they're weird in general. Key Piece can be lit up either when any Key Piece is obtained, or when you have enough Key Piece items to pass the threshold, whichever the dev decides.